+++ content.js
@@ -8,23 +8,32 @@ async function checkForDID(domain) {
     `https://dns.google/resolve?name=_atproto.${domain}&type=TXT`
   )
   const data = await response.json()
+
   const records = data?.Answer?.filter((record) => record.type === 16) || []
-  return records.some((record) => record.data.includes("did=did:plc:"))
+
+  const didRecord = records.find((record) =>
+    record.data.includes("did=did:plc:")
+  )
+
+  return didRecord ? didRecord.data.replace("did=", "") : null
 }
 
 ;(async function () {
   const domain = getDomainName()
-  const didFound = await checkForDID(domain)
+  const did = await checkForDID(domain)
 
-  if (didFound) {
-    chrome.runtime.sendMessage({ type: "DID_FOUND" })
+  if (did) {
+    chrome.runtime.sendMessage({ type: "DID_FOUND", did })
   } else {
     chrome.runtime.sendMessage({ type: "DID_NOT_FOUND" })
   }
 })()
 
 chrome.runtime.onMessage.addListener((message, sender, sendResponse) => {
-  if (message.type === "GET_DOMAIN") {
-    sendResponse({ domain: getDomainName() })
+  if (message.type === "GET_DID") {
+    checkForDID(getDomainName())
+      .then((did) => sendResponse({ did }))
+      .catch(() => sendResponse({ did: null }))
+    return true // Indicate that the response will be sent asynchronously.
   }
 })